Cosmetics... again.
Some 'align' attributes added in interfaces.

In download.xul, I changed 'maxthreads' and 'numretries' max value (same as in options.xul)

In options.xul, the 'disabled' attribute is no longer in the persist attribute.
I changed the code to set 'disabled' on loading in the 'setAccBut()' function.
There was an issue with 'Cancel' button when the 'disable' attribute was changed.

Signed-off-by: Rookie <Thierry@.(none)>