Host Timeout between downloads becomes a Host attribute.
Modification of the HostFile Editor in order to introduce the Timeout in the host definition.
Other hosts (even with max. threads limit) will not be affected by the global Timeout Between Downloads parameter anymore.
The code to change this behaviour is still under progress, this commit just changes the HostFile Editor.
To evaluate the modification, change the line in hostfile_ui.xul as 'commented'.

Signed-off-by: Rookie <Thierry@.(none)>