Thread-sucker missing pages.
Some forums redirect page requests to sponsor/partners pages.
When the redirection concerns the image grabbing process,
the D/L window offers a second chance with the action "Retry".
For the pages collected by the threadsucker, the redirection means
the loss of the images for this page. The collection is amputated.

FIX : IHG thread-sucker detects now that the final URI of the requested
thread pages is not the expected one and resends the XHR with the original URI.

Signed-off-by: Rookie <Rookie@gmail.com>